pyproject.toml
[project]
name = "xabironelson-codex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "typer>=0.19.2",
]
